modified:   progs/ngram_counts.cpp 		added debug messages for thread-based counts
	modified:   progs/ngram_counts_mpi.cpp
		I'm surprised this code basically works...!
		bugfix for help messages.
		bugfix for counts removal after collection...
		added debug messages
		added --prog option to make it compatible with other mpi-based applications

	modified:   utils/byte_aligned_code.hpp
		casting into uint16_t

+++ utils/byte_aligned_code.hpp
@@ -39,7 +39,7 @@ namespace utils
     size_t byte_size(Tp x)  { return __byte_size((uint16_t) x); }
     
     static const uint16_t mask1 = ~uint16_t(0x7f);
-    static const uint16_t mask2 = mask1 << 7;
+    static const uint16_t mask2 = uint16_t(mask1 << 7);
 
     static inline
     size_t __byte_size(uint16_t x)
